OT 17.1.0.10447

+++ include/comm/net.cpp
@@ -46,10 +46,8 @@
 #endif
 
 #define WIN32_LEAN_AND_MEAN
-//#include <windows.h>
 
 #include <winsock.h>
-#include <stdarg.h>
 #include <process.h>
 
 
@@ -578,7 +576,7 @@ namespace coid {
 
 #ifdef SYSTYPE_WIN
 
-        uint32 nblocking = blocking ? 0 : 1;
+        ulong nblocking = blocking ? 0 : 1;
         ::ioctlsocket(handle, FIONBIO, &nblocking);
 
 #else
